\name{retail}
\alias{retail}
\docType{data}
\title{
    dataset retail
}
\description{
    some description of the dataset
}
\format{
    A data frame with 228 observations on 62 variables.
}
\source{
    optional: reference to a publication or URL from which the data were obtained
}

\examples{
data(retail)
str(retail)
retail$AutomobileDealers
}
