\name{xm_sample1}
\alias{xm_sample1}
\docType{data}
\title{
    dataset xm_sample1
}
\description{
    some description of the dataset
}
\format{
    A data frame with 167 observations on the following 35 variables.
}
\source{
    optional: reference to a publication or URL from which the data were obtained
}

\examples{
data(xm_sample1)
str(xm_sample1)
}
