\name{retail}
\alias{retail}
\docType{data}
\title{
  US-retail
}
\description{
  US-retail
}
\usage{data("retail")}
\format{
  A data frame with 228 observations on the following 62 variables.
  \describe{
    \item{\code{AllOtherGenMerchandiseStores}}{a matrix with 1 columns}
    \item{\code{AllOtherHomeFurnishingsStores}}{a matrix with 1 columns}
    \item{\code{AppltvAndOtherElectStores}}{a matrix with 1 columns}
    \item{\code{AutomobileDealers}}{a matrix with 1 columns}
    \item{\code{BeerWineAndLiquorStores}}{a matrix with 1 columns}
    \item{\code{BookStores}}{a matrix with 1 columns}
    \item{\code{BuildingMatAndGardenEquipAndSupp}}{a matrix with 1 columns}
    \item{\code{BuildingMatAndSuppliesDealers}}{a matrix with 1 columns}
    \item{\code{ClothingAndClothingAccessStores}}{a matrix with 1 columns}
    \item{\code{ClothingStores}}{a matrix with 1 columns}
    \item{\code{ComputerAndSoftwareStores}}{a matrix with 1 columns}
    \item{\code{DepartmentStoresExclDiscountDepa}}{a matrix with 1 columns}
    \item{\code{DepartmentStoresExclLD}}{a matrix with 1 columns}
    \item{\code{DepartmentStoresInclLD}}{a matrix with 1 columns}
    \item{\code{DiscountDeptStoresInclLD}}{a matrix with 1 columns}
    \item{\code{DiscountDeptStores}}{a matrix with 1 columns}
    \item{\code{DrinkingPlaces}}{a matrix with 1 columns}
    \item{\code{ElectronicsAndApplianceStores}}{a matrix with 1 columns}
    \item{\code{FamilyClothingStores}}{a matrix with 1 columns}
    \item{\code{FloorCoveringStores}}{a matrix with 1 columns}
    \item{\code{FoodAndBeverageStores}}{a matrix with 1 columns}
    \item{\code{FoodServicesAndDrinkingPlaces}}{a matrix with 1 columns}
    \item{\code{FuelDealers}}{a matrix with 1 columns}
    \item{\code{FullServiceRestaurants}}{a matrix with 1 columns}
    \item{\code{FurnitureAndHomeFurnishingsStore}}{a matrix with 1 columns}
    \item{\code{FurnitureHomeFurnElectronicsAndA}}{a matrix with 1 columns}
    \item{\code{FurnitureStores}}{a matrix with 1 columns}
    \item{\code{Gafo}}{a matrix with 1 columns}
    \item{\code{GasolineStations}}{a matrix with 1 columns}
    \item{\code{GeneralMerchandiseStores}}{a matrix with 1 columns}
    \item{\code{GiftNoveltyAndSouvenirStores}}{a matrix with 1 columns}
    \item{\code{GroceryStores}}{a matrix with 1 columns}
    \item{\code{HardwareStores}}{a matrix with 1 columns}
    \item{\code{HealthAndPersonalCareStores}}{a matrix with 1 columns}
    \item{\code{HobbyToyAndGameStores}}{a matrix with 1 columns}
    \item{\code{HomeFurnishingsStores}}{a matrix with 1 columns}
    \item{\code{HouseholdApplianceStores}}{a matrix with 1 columns}
    \item{\code{JewelryStores}}{a matrix with 1 columns}
    \item{\code{LimitedServiceEatingPlaces}}{a matrix with 1 columns}
    \item{\code{MensClothingStores}}{a matrix with 1 columns}
    \item{\code{MiscellaneousStoreRetailers}}{a matrix with 1 columns}
    \item{\code{MotorVehicleAndPartsDealers}}{a matrix with 1 columns}
    \item{\code{NewCarDealers}}{a matrix with 1 columns}
    \item{\code{NonstoreRetailers}}{a matrix with 1 columns}
    \item{\code{OfficeSuppliesAndStationeryStore}}{a matrix with 1 columns}
    \item{\code{OfficeSuppliesStationeryAndGiftS}}{a matrix with 1 columns}
    \item{\code{OtherClothingStores}}{a matrix with 1 columns}
    \item{\code{OtherGeneralMerchandiseStores}}{a matrix with 1 columns}
    \item{\code{PaintAndWallpaperStores}}{a matrix with 1 columns}
    \item{\code{PharmaciesAndDrugStores}}{a matrix with 1 columns}
    \item{\code{RadioTVAndOtherElectStores}}{a matrix with 1 columns}
    \item{\code{RetailAndFoodServicesSalesTotal}}{a matrix with 1 columns}
    \item{\code{RetailSalesTotalExclMotorVehicle}}{a matrix with 1 columns}
    \item{\code{RetailSalesTotal}}{a matrix with 1 columns}
    \item{\code{ShoeStores}}{a matrix with 1 columns}
    \item{\code{SportingGoodsHobbyBookAndMusicSt}}{a matrix with 1 columns}
    \item{\code{SportingGoodsStores}}{a matrix with 1 columns}
    \item{\code{SupermarketsAndOtherGroceryExcep}}{a matrix with 1 columns}
    \item{\code{UsedCarDealers}}{a matrix with 1 columns}
    \item{\code{UsedMerchandiseStores}}{a matrix with 1 columns}
    \item{\code{WarehouseClubsAndSuperstores}}{a matrix with 1 columns}
    \item{\code{WomensClothingStores}}{a matrix with 1 columns}
  }
}
\details{
}
\source{
}
\references{
}
\examples{
data(retail)
}
\keyword{datasets}
